Additional bug fixes and correct hmmer index reporting in json. As well as bug fixes to the vectorisation of coordinates such that it isn't tripped up by pandas indices.
Updated dtype issue with coordinates conversion to pass tests. Removed bad fetch nc file present in screeners.

Removed erroneous print debug statements.

+++ commec/utils/coordinates.py
@@ -24,16 +24,17 @@ def convert_protein_to_nucleotide_coords(frame,
     tuple: (nucleotide_start, nucleotide_end)
     """
     # Convert protein coordinates to 0-based, for calculation.
-    protein_start = np.asarray(protein_start) - 1
-    protein_end = np.asarray(protein_end) - 1
-    frame = np.asarray(frame)
-    seq_length = np.asarray(seq_length)
+    protein_start = np.asarray(protein_start, dtype=np.int64) - 1
+    protein_end = np.asarray(protein_end, dtype=np.int64) - 1
+    frame = np.asarray(frame, dtype=np.int64)
+    seq_length = np.asarray(seq_length, dtype=np.int64)
 
+    # Reverse frame offsets, for when total length not divisible into codons.
     reverse_offset = seq_length % 3
 
     # Initialize arrays for nucleotide start and end
-    nucleotide_start = np.zeros_like(protein_start, dtype=np.int32)
-    nucleotide_end = np.zeros_like(protein_end, dtype=np.int32)
+    nucleotide_start = np.zeros_like(protein_start, dtype=np.int64)
+    nucleotide_end = np.zeros_like(protein_end, dtype=np.int64)
 
     # Forward frames (1, 2, 3)
     forward_mask = frame <= 3
@@ -48,13 +49,6 @@ def convert_protein_to_nucleotide_coords(frame,
 
     nucleotide_start[reverse_mask] = seq_length[reverse_mask] - nuc_end_reverse - 1 - reverse_offset[reverse_mask]
     nucleotide_end[reverse_mask] = seq_length[reverse_mask] - nuc_start_reverse - 1 - reverse_offset[reverse_mask]
-
-    # Frames 2,3, 5,6, have 1 less qlen, thus require a codon offset. But only for reverse frames.
-    # We are now using exact nt_qlen, so this may no longer be necessary...
-    #  ... but we MAY need offsets depending on %3 the length.
-    # qlen_reverse_mask = frame > 4
-    # nucleotide_start[qlen_reverse_mask] += 3
-    # nucleotide_end[qlen_reverse_mask] += 3
 
     # Convert to back to 1-based coordinates for reporting.
     nucleotide_start += 1
